feat: Support Souin as Caddy plugin
* Move Souin legacy as plugin
* Remove main package declaration
* Move header XFromCache setter
