+++ Makefile
@@ -8,6 +8,9 @@ build-app: env-prod ## Build containers with prod env vars
 	$(DC_BUILD) souin
 	$(MAKE) up
 
+build-and-run-caddy: ## Build caddy binary
+	cd plugins/caddy && xcaddy build --with github.com/darkweak/souin/plugins/caddy=./ --with github.com/darkweak/souin@latest=../.. && ./caddy run
+
 build-dev: env-dev ## Build containers with dev env vars
 	$(DC_BUILD) souin
 	$(MAKE) up
