+++ main.go
@@ -1,7 +1,1 @@
-package main
-
-import "github.com/darkweak/souin/cache"
-
-func main() {
-	cache.Start()
-}
+package souin
